Review of the fuse mount chain of command
**Tracing**
* storage/gcs: added logging and debug traces

* cmd(CLI): refactored logger injection, added more commands subject to log level setting

* core: enabled fuse internal debug logger (experimental, since it is quite verbose)

* cafs: injected logger

**Tuning/user experimentation**
* cmd(CLI): added new options to tune fuse mount cache size and prefetch behavior (e.g. --cache-size 100MB, --prefetch 3)

* core: bundle conveys new options down to cafs factory

**Refactoring & readability**
* core: refactored NewXXXFS to suppress redundant staging path and logging (following initial idea from PR#354)
* core: refactored RO mount for readability

* cafs: more godoc, resolved question marks left in code
* cafs: refactoring of struct factories, chunkReader and writer, more options to control LRU cache and prefetch behavior
* cafs: refactored hasher ("helpers") for readability

**Testing**
* cmd(CLI): re-enacted fuse CLI e2e testing (runs on CI)

* cafs: added more unit tests on freelists, hasher and ReadAt with concurrency testing

**Fixes**
* core: fixed spurious warnings on fs (extended attributes, flush on RO mount)

* cafs: enabled evicted LRU buffers to be relinquished to freelist
* cafs: fixed concurrency issues with corrupted relinquished buffers
* cafs: fixed leaf hash verification for ReadAt

**Performance improvements**
* cafs: set up cache for resolved leaf keys from root
* cafs: more efficient prefetching, with much less waste of work and prefetching blobs with tunable look-ahead

**Memory footprint reduction**
* core: relinquished unused RadixTree after ro fs construction

* cafs: adapted freelist buffers to leaf size, not systematically max size

**CI changes**
* cafs: cafs tests are memory hungry: run them in CI separately with different settings (less parallelism, no race)
* hack/fuse_demo: adapted fuse mount detection, since we slightly changed the mount signature (see below)

**Other enhancements**
* core: set fuse moumt subtype to "datamon" and "datamon-mutable", so it appears as such in the output of the "mount" UNIX command

+++ cmd/metrics/cmd/write_files.go
@@ -68,7 +68,7 @@ yet is generally part of the datamon metrics collection and benchmarking picture
 				filenames = append(filenames, nextFileName)
 			}
 			max := int64(1024 * 1024 * params.writeFiles.fileSize)
-			s = newGenStoreZeroOneChunks(filenames, max, cafs.DefaultLeafSize)
+			s = newGenStoreZeroOneChunks(filenames, max, int64(cafs.DefaultLeafSize))
 			return s
 		}()
